Refactored meta generator
