+++ metagenerator/api.go
@@ -1,0 +1,87 @@
+package metagenerator
+
+import (
+	"bytes"
+	"encoding/json"
+	"fmt"
+	"io"
+	"net/http"
+	"net/http/httputil"
+)
+
+type Query struct {
+	Path  string `json:"path"`
+	Query string `json:"query"`
+}
+
+func putMeta(record *Record, apiKey, projectId, url string) error {
+	req, err := json.Marshal(record)
+	if err != nil {
+		return err
+	}
+
+	r, err := http.NewRequest("PUT", url, bytes.NewBuffer(req))
+	if err != nil {
+		return err
+	}
+	r.Header.Add("Content-Type", "application/json")
+	r.Header.Add("Authorization", fmt.Sprintf("Bearer %s", apiKey))
+	r.Header.Add("X-Project-ID", projectId)
+
+	/* TODO: add back with log levels
+	re, err := httputil.DumpRequest(r, true)
+	if err != nil {
+		return err
+	}
+	fmt.Println(string(re))
+	*/
+
+	client := &http.Client{}
+	res, err := client.Do(r)
+	if err != nil {
+		return err
+	}
+	defer res.Body.Close()
+
+	if res.StatusCode != http.StatusOK {
+		return err
+	}
+
+	return nil
+}
+
+func SearchMeta(query Query, apiKey, projectId, url string) error {
+	req, err := json.Marshal(query)
+	if err != nil {
+		return err
+	}
+	r, err := http.NewRequest("POST", url, bytes.NewBuffer(req))
+	if err != nil {
+		return err
+	}
+	r.Header.Add("Content-Type", "application/json")
+	r.Header.Add("Authorization", fmt.Sprintf("Bearer %s", apiKey))
+	r.Header.Add("X-Project-ID", projectId)
+
+	reqest, err := httputil.DumpRequest(r, true)
+	if err != nil {
+		return err
+	}
+	fmt.Println(string(reqest))
+
+	client := &http.Client{}
+	res, err := client.Do(r)
+	if err != nil {
+		return err
+	}
+	defer res.Body.Close()
+
+	bodyBytes, err := io.ReadAll(res.Body)
+	if err != nil {
+		panic(err)
+	}
+	bodyString := string(bodyBytes)
+	fmt.Println(bodyString)
+
+	return nil
+}
